Factory design pattern on Location
