+++ app/coordinates.py
@@ -16,3 +16,22 @@ class Point:
 def __init__(self, x, y):  
 self.x = x  
 self.y = y  
+
+
+
+class Person: 
+
+def __init__(self, p, w):
+	self.p = p
+	self.w = w
+
+personOne = Person("Abdullah", 190)
+
+print(personOne.p)
+print(personOne.w)
+
+3
+y
+n
+y
+y
