+++ app/location/DataLocation
@@ -1,0 +1,53 @@
+from ..coordinates import Coordinates
+from ..utils import countries
+from ..utils.populations import country_population
+from __init__ import country_population, country_code
+
+# pylint: disable=redefined-builtin,invalid-name
+class Location:  # pylint: disable=too-many-instance-attributes
+    """
+    A location in the world affected by the coronavirus.
+    """
+
+    def __init__(
+        self, id, country, province, coordinates, last_updated, confirmed, deaths, recovered,
+    ):  # pylint: disable=too-many-arguments
+        # General info.
+        self.id = id
+        self.country = country.strip()
+        self.province = province.strip()
+        self.coordinates = coordinates
+
+        # Last update.
+        self.last_updated = last_updated
+
+        # Statistics.
+        self.confirmed = confirmed
+        self.deaths = deaths
+        self.recovered = recovered
+
+         def serialize(self):
+        """
+        Serializes the location into a dict.
+
+        :returns: The serialized location.
+        :rtype: dict
+        """
+        return {
+            # General info.
+            "id": self.id,
+            "country": self.country,
+            "country_code": self.country_code,
+            "country_population": self.country_population,
+            "province": self.province,
+            # Coordinates.
+            "coordinates": self.coordinates.serialize(),
+            # Last updated.
+            "last_updated": self.last_updated,
+            # Latest data (statistics).
+            "latest": {
+                "confirmed": self.confirmed,
+                "deaths": self.deaths,
+                "recovered": self.recovered,
+            },
+        }
