+++ app/location/__init__.py
@@ -10,23 +10,6 @@ class Location:  # pylint: disable=too-many-instance-attributes
     A location in the world affected by the coronavirus.
     """
 
-    def __init__(
-        self, id, country, province, coordinates, last_updated, confirmed, deaths, recovered,
-    ):  # pylint: disable=too-many-arguments
-        # General info.
-        self.id = id
-        self.country = country.strip()
-        self.province = province.strip()
-        self.coordinates = coordinates
-
-        # Last update.
-        self.last_updated = last_updated
-
-        # Statistics.
-        self.confirmed = confirmed
-        self.deaths = deaths
-        self.recovered = recovered
-
     @property
     def country_code(self):
         """
@@ -46,33 +29,6 @@ class Location:  # pylint: disable=too-many-instance-attributes
         :rtype: int
         """
         return country_population(self.country_code)
-
-    def serialize(self):
-        """
-        Serializes the location into a dict.
-
-        :returns: The serialized location.
-        :rtype: dict
-        """
-        return {
-            # General info.
-            "id": self.id,
-            "country": self.country,
-            "country_code": self.country_code,
-            "country_population": self.country_population,
-            "province": self.province,
-            # Coordinates.
-            "coordinates": self.coordinates.serialize(),
-            # Last updated.
-            "last_updated": self.last_updated,
-            # Latest data (statistics).
-            "latest": {
-                "confirmed": self.confirmed,
-                "deaths": self.deaths,
-                "recovered": self.recovered,
-            },
-        }
-
 
 class TimelinedLocation(Location):
     """
